requirements.txt
youtube_dl
python-cfonts